feat: alterado banco de dados para postgres. adicionado módulo de biblioteca. refatorado schema de membros

+++ src/membros/membros.controller.ts
@@ -1,11 +1,11 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, HttpCode } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, HttpCode, ParseUUIDPipe, HttpException, HttpStatus } from '@nestjs/common';
 import { MembrosService } from './membros.service';
 import { CreateMembroDto } from './dto/create-membro.dto';
 import { UpdateMembroDto } from './dto/update-membro.dto';
 import { ApiTags } from '@nestjs/swagger';
 
-@Controller('membros')
-@ApiTags('membros')
+@Controller('membro')
+@ApiTags('membro')
 export class MembrosController {
   constructor(private readonly membrosService: MembrosService) {}
 
@@ -19,20 +19,33 @@ export class MembrosController {
     return this.membrosService.findAll();
   }
 
+  // must be UUID
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.membrosService.findOne(id);
+  async findOne(@Param('id', ParseUUIDPipe) id: string) {
+    const membro = await this.membrosService.findOne(id);
+    if (!membro) {
+      throw new HttpException('Membro não encontrado', HttpStatus.NOT_FOUND);
+    }
+    return membro;
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateMembroDto: UpdateMembroDto) {
-    return this.membrosService.update(id, updateMembroDto);
+  async update(@Param('id') id: string, @Body() updateMembroDto: UpdateMembroDto) {
+    const membro = await this.membrosService.findOne(id);
+    if (!membro) {
+      throw new HttpException('Membro não encontrado', HttpStatus.NOT_FOUND);
+    }
+    return await this.membrosService.update(id, updateMembroDto);
   }
 
   @Delete(':id')
   @HttpCode(204)
-  remove(@Param('id') id: string) {
-    this.membrosService.remove(id);
+  async remove(@Param('id', ParseUUIDPipe) id: string) {
+    const membro = await this.membrosService.findOne(id);
+    if (!membro) {
+      throw new HttpException('Membro não encontrado', HttpStatus.NOT_FOUND);
+    }
+    await this.membrosService.remove(id);
     return null;
   }
 }
